requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.42.0
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-telegram-bot==21.4
pyzbar==0.1.9
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
